feat: create stories.
