+++ src/components/Layout/Router.tsx
@@ -4,8 +4,7 @@ import NavBar from '@components/NavBar/NavBar'
 import { SidebarNav } from '@components/SidebarNav/SidebarNav'
 import React from 'react'
 import EmptyProjectPage from '@pages/Login/EmptyProjectPage'
-import { ProjectAddition } from '@components/ProjectAddition/ProjectAddition'
-import WrapperPage from '@components/Wrapper/WrapperPage'
+import LoginPage from '@pages/Login/LoginPage'
 
 const WithNav = (): JSX.Element => {
   const [isMobile, setIsMobile] = React.useState<boolean>(
@@ -39,50 +38,11 @@ const WithNav = (): JSX.Element => {
     </div>
   )
 }
-const projects = [
-  {
-    id: '1',
-    name: 'Project 1',
-    url: '',
-    createdAt: new Date(),
-    usersRoles: [],
-    pendingUsers: [],
-    projectStages: [],
-    issues: []
-  },
-  {
-    id: '2',
-    name: 'Project 2',
-    url: '',
-    createdAt: new Date(),
-    usersRoles: [],
-    pendingUsers: [],
-    projectStages: [],
-    issues: []
-  },
-  {
-    id: '3',
-    name: 'Project 3',
-    url: '',
-    createdAt: new Date(),
-    usersRoles: [],
-    pendingUsers: [],
-    projectStages: [],
-    issues: []
-  }
-]
+
 export const ROUTER = createBrowserRouter([
   {
     path: '/login',
-    element: (
-      <WrapperPage>
-        <ProjectAddition
-          handleToken={() => {}}
-          handleSelectedProject={() => {}}
-          projects={projects}
-        />
-      </WrapperPage>
-    )
+    element: <LoginPage /> // StepperExample
   },
   {
     path: '/login/non-invited',
